Hi, I’m @CyberFaux2pointOh
Well as I am getting better and more into debugging/patching I have come to a sudden epiphany! As as former victim of ID theft, I have now made it my mission to do anything possible to prevent it from happening to anyone else. I want to create a bunch of different apps to help prevent it upfront and personal (such as preventing android beam, etc.) I have noticed that some of the technology we currently use is outdated as cyber criminals mindsets are constantly changing and evolving around our current laws. I have the will and tenacity do to everything possible to help prevent it. If you are in the field of cyber security please email me and let me know as I am about to head to school and learn more about it! thank you very much!!!